# tests/snapshots/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BriefClaimModel": {
        "additionalProperties": false,
        "description": "One claim in the brief, with its support level.",
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "included": {
            "title": "Included",
            "type": "boolean"
          },
          "source_numbers": {
            "items": {
              "type": "integer"
            },
            "title": "Source Numbers",
            "type": "array"
          },
          "support": {
            "title": "Support",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "text",
          "support",
          "source_numbers",
          "included"
        ],
        "title": "BriefClaimModel",
        "type": "object"
      },
      "BriefNotesRequest": {
        "additionalProperties": false,
        "description": "Replace the patient's own notes.",
        "properties": {
          "notes": {
            "default": "",
            "maxLength": 4000,
            "title": "Notes",
            "type": "string"
          }
        },
        "title": "BriefNotesRequest",
        "type": "object"
      },
      "BriefQuestionModel": {
        "additionalProperties": false,
        "description": "One question in the brief, with where it came from.",
        "properties": {
          "origin": {
            "description": "Whether the patient or the system wrote it.",
            "title": "Origin",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "question_id",
          "text",
          "origin"
        ],
        "title": "BriefQuestionModel",
        "type": "object"
      },
      "BriefQuestionOrderRequest": {
        "additionalProperties": false,
        "description": "Reorder the questions by identifier.",
        "properties": {
          "order": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 0,
            "title": "Order",
            "type": "array"
          }
        },
        "required": [
          "order"
        ],
        "title": "BriefQuestionOrderRequest",
        "type": "object"
      },
      "BriefQuestionRequest": {
        "additionalProperties": false,
        "description": "Add one question the patient wrote themselves.",
        "properties": {
          "text": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "BriefQuestionRequest",
        "type": "object"
      },
      "BriefResponse": {
        "additionalProperties": false,
        "description": "The server-owned brief. A client edits it by named operation only.",
        "properties": {
          "available_sections": {
            "items": {
              "type": "string"
            },
            "title": "Available Sections",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/BriefClaimModel"
            },
            "title": "Claims",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "export_warning": {
            "title": "Export Warning",
            "type": "string"
          },
          "fill_ratio": {
            "title": "Fill Ratio",
            "type": "number"
          },
          "fits_one_page": {
            "title": "Fits One Page",
            "type": "boolean"
          },
          "included_sections": {
            "items": {
              "type": "string"
            },
            "title": "Included Sections",
            "type": "array"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "overflow_advice": {
            "items": {
              "type": "string"
            },
            "title": "Overflow Advice",
            "type": "array"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/BriefQuestionModel"
            },
            "title": "Questions",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceModel"
            },
            "title": "Sources",
            "type": "array"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "turn_index": {
            "title": "Turn Index",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "turn_index",
          "topic",
          "notes",
          "questions",
          "claims",
          "sources",
          "included_sections",
          "available_sections",
          "fits_one_page",
          "fill_ratio",
          "overflow_advice",
          "export_warning",
          "disclaimer"
        ],
        "title": "BriefResponse",
        "type": "object"
      },
      "BriefSectionsRequest": {
        "additionalProperties": false,
        "description": "Choose which sections the export includes.",
        "properties": {
          "sections": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "minItems": 0,
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "sections"
        ],
        "title": "BriefSectionsRequest",
        "type": "object"
      },
      "BriefTopicRequest": {
        "additionalProperties": false,
        "description": "Replace the patient's own 'what I want to talk about' line.",
        "properties": {
          "topic": {
            "default": "",
            "maxLength": 2000,
            "title": "Topic",
            "type": "string"
          }
        },
        "title": "BriefTopicRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceModel": {
        "additionalProperties": false,
        "description": "One numbered source, as it appears inline and in the source panel.",
        "properties": {
          "number": {
            "description": "Stable display number, assigned from retrieval order.",
            "minimum": 1.0,
            "title": "Number",
            "type": "integer"
          },
          "relevance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Retrieval relevance, 0..1. NOT a confidence and NOT a quality grade.",
            "title": "Relevance"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "description": "Empty when the recorded link was not http(s).",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "number",
          "source_id",
          "title",
          "url"
        ],
        "title": "SourceModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend for Synapse, a pre-clinical appointment-preparation prototype.\n\n**Not a medical device.** Not validated, not FDA cleared, not HIPAA compliant.\nNo clinician has reviewed any source, evaluation case, threshold or output.\n\nEvery route except `/healthz` and `/readyz` requires the `X-Service-Token`\nheader. Conversation routes additionally require an access token obtained from\n`/v1/access/login`. The browser is never expected to hold the service token: it\ncalls a same-origin server proxy, which forwards.\n",
    "title": "Synapse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/turns/{turn_index}/brief": {
      "get": {
        "description": "Build the brief on first access, then return the stored one.",
        "operationId": "read_brief_v1_turns__turn_index__brief_get",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the brief",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/export/{fmt}": {
      "get": {
        "description": "Render the brief and return it as an attachment.\n\nBuilt in memory and streamed out. Nothing touches disk.",
        "operationId": "export_brief_v1_turns__turn_index__brief_export__fmt__get",
        "parameters": [
          {
            "in": "path",
            "name": "fmt",
            "required": true,
            "schema": {
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the brief",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/notes": {
      "put": {
        "description": "The patient's own notes.",
        "operationId": "put_notes_v1_turns__turn_index__brief_notes_put",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set the notes",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/questions": {
      "post": {
        "description": "Add a question the patient wrote. Marked as theirs, never as evidence.",
        "operationId": "post_question_v1_turns__turn_index__brief_questions_post",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a question",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/questions/order": {
      "put": {
        "description": "Reorder by identifier. The set must match exactly; nothing is added here.",
        "operationId": "put_question_order_v1_turns__turn_index__brief_questions_order_put",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefQuestionOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder questions",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/questions/{question_id}": {
      "delete": {
        "description": "Remove one question by identifier.",
        "operationId": "delete_question_v1_turns__turn_index__brief_questions__question_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          },
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove a question",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/sections": {
      "put": {
        "description": "Choose which sections the export includes.\n\nThe disclaimer is always included and cannot be removed — the section list\nthe schema exposes does not contain it.",
        "operationId": "put_sections_v1_turns__turn_index__brief_sections_put",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefSectionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose sections",
        "tags": [
          "brief"
        ]
      }
    },
    "/v1/turns/{turn_index}/brief/topic": {
      "put": {
        "description": "The patient's own 'what I want to talk about' line.",
        "operationId": "put_topic_v1_turns__turn_index__brief_topic_put",
        "parameters": [
          {
            "description": "Zero-based index of the turn.",
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "description": "Zero-based index of the turn.",
              "maximum": 1000,
              "minimum": 0,
              "title": "Turn Index",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefTopicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set the topic",
        "tags": [
          "brief"
        ]
      }
    }
  }
}